requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
requests>=2.27.1
requests-oauthlib>=1.3.1